moved pytest.fixtures to conftest fixted yamllint for running locally fixed yamllint errors
