+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
       - id: check-yaml
       - id: check-ast
       - id: name-tests-test
-        args: ["--django"] # ensure test files begin 'test_'
+        args: ["--django"]  # ensure test files begin 'test_'
   - repo: https://github.com/ambv/black
     rev: 22.3.0
     hooks:
