+++ tests/test_reader_api.py
@@ -73,30 +73,6 @@ invalid_file_type_engine_combinations = [
 invalid_engines = ["spark", "dplyr"]
 
 
-@pytest.fixture
-def df_all_types():
-    return pd.read_csv("tests/data/all_types.csv")
-
-
-@pytest.fixture
-def test_meta():
-    return {
-        "columns": [
-            {"name": "my_float", "type": "float64", "type_category": "float"},
-            {"name": "my_bool", "type": "bool_", "type_category": "boolean"},
-            {"name": "my_nullable_bool", "type": "bool_", "type_category": "boolean"},
-            {"name": "my_date", "type": "date32", "type_category": "timestamp"},
-            {
-                "name": "my_datetime",
-                "type": "timestamp(s)",
-                "type_category": "timestamp",
-            },
-            {"name": "my_int", "type": "int64", "type_category": "integer"},
-            {"name": "my_string", "type": "string", "type_category": "string"},
-        ]
-    }
-
-
 @pytest.mark.parametrize("data_format", ["jsonl", "csv"])
 def test_inferred_cols_pandas_types(data_format):
     df = reader.read(f"tests/data/all_types.{data_format}")
