+++ tests/test_round_trip.py
@@ -37,30 +37,6 @@ valid_reader_engines = valid_engines
 valid_writer_engines = valid_engines
 
 
-@pytest.fixture
-def test_meta():
-    return {
-        "columns": [
-            {"name": "my_float", "type": "float64", "type_category": "float"},
-            {"name": "my_bool", "type": "bool_", "type_category": "boolean"},
-            {"name": "my_nullable_bool", "type": "bool_", "type_category": "boolean"},
-            {"name": "my_date", "type": "date32", "type_category": "timestamp"},
-            {
-                "name": "my_datetime",
-                "type": "timestamp(s)",
-                "type_category": "timestamp",
-            },
-            {"name": "my_int", "type": "int64", "type_category": "integer"},
-            {"name": "my_string", "type": "string", "type_category": "string"},
-        ]
-    }
-
-
-@pytest.fixture
-def df_all_types_from_meta(test_meta):
-    return reader.csv.read("tests/data/all_types.csv", test_meta)
-
-
 @pytest.mark.parametrize("trip1_file_format", all_formats)
 @pytest.mark.parametrize("trip2_file_format", all_formats)
 def test_round_trip(
